move Context modules in a directory
